Move tests to test_path
