+++ tests/test_activate.py
@@ -34,7 +34,6 @@ from conda.activate import (
     _build_activator_cls,
     activator_map,
     native_path_to_unix,
-    unix_path_to_native,
 )
 from conda.auxlib.ish import dals
 from conda.base.constants import (
@@ -1106,294 +1105,6 @@ def make_dot_d_files(prefix, extension):
 
     touch(join(prefix, "etc", "conda", "activate.d", f"activate1{extension}"))
     touch(join(prefix, "etc", "conda", "deactivate.d", f"deactivate1{extension}"))
-
-
-@pytest.mark.skipif(
-    not on_win,
-    reason="native_path_to_unix is path_identity on non-windows",
-)
-@pytest.mark.parametrize(
-    "paths,expected",
-    [
-        # falsy
-        pytest.param(None, [None], id="None"),
-        pytest.param("", ["."], id="empty string"),
-        pytest.param((), [()], id="empty tuple"),
-        # native
-        pytest.param(
-            "C:\\path\\to\\One",
-            [
-                "/c/path/to/One",  # MSYS2
-                "/cygdrive/c/path/to/One",  # cygwin
-            ],
-            id="path",
-        ),
-        pytest.param(
-            ["C:\\path\\to\\One"],
-            [
-                ("/c/path/to/One",),  # MSYS2
-                ("/cygdrive/c/path/to/One",),  # cygwin
-            ],
-            id="list[path]",
-        ),
-        pytest.param(
-            ("C:\\path\\to\\One", "C:\\path\\Two", "\\\\mount\\Three"),
-            [
-                ("/c/path/to/One", "/c/path/Two", "//mount/Three"),  # MSYS2
-                (
-                    "/cygdrive/c/path/to/One",
-                    "/cygdrive/c/path/Two",
-                    "//mount/Three",
-                ),  # cygwin
-            ],
-            id="tuple[path, ...]",
-        ),
-        pytest.param(
-            "C:\\path\\to\\One;C:\\path\\Two;\\\\mount\\Three",
-            [
-                "/c/path/to/One:/c/path/Two://mount/Three",  # MSYS2
-                "/cygdrive/c/path/to/One:/cygdrive/c/path/Two://mount/Three",  # cygwin
-            ],
-            id="path;...",
-        ),
-    ],
-)
-@pytest.mark.parametrize(
-    "cygpath",
-    [pytest.param(True, id="cygpath"), pytest.param(False, id="fallback")],
-)
-def test_native_path_to_unix(
-    mocker: MockerFixture,
-    paths: str | Iterable[str] | None,
-    expected: str | list[str] | None,
-    cygpath: bool,
-) -> None:
-    if not cygpath:
-        # test without cygpath
-        mocker.patch("subprocess.run", side_effect=FileNotFoundError)
-
-    assert native_path_to_unix(paths) in expected
-
-
-@pytest.mark.skipif(
-    not on_win,
-    reason="native_path_to_unix is path_identity on non-windows",
-)
-@pytest.mark.parametrize(
-    "paths,expected",
-    [
-        # falsy
-        pytest.param(None, None, id="None"),
-        pytest.param("", ".", id="empty string"),
-        pytest.param((), (), id="empty tuple"),
-        # MSYS2
-        pytest.param(
-            # 1 leading slash = root
-            "/",
-            "{WINDOWS}\\Library\\",
-            id="root",
-        ),
-        pytest.param(
-            # 1 leading slash + 1 letter = drive
-            "/c",
-            "C:\\",
-            id="drive",
-        ),
-        pytest.param(
-            # 1 leading slash + 1 letter = drive
-            "/c/",
-            "C:\\",
-            id="drive [trailing]",
-        ),
-        pytest.param(
-            # 1 leading slash + 2+ letters = root path
-            "/root",
-            "{WINDOWS}\\Library\\root",
-            id="root path",
-        ),
-        pytest.param(
-            # 1 leading slash + 2+ letters = root path
-            "/root/",
-            "{WINDOWS}\\Library\\root\\",
-            id="root path [trailing]",
-        ),
-        pytest.param(
-            # 2 leading slashes = UNC mount
-            "//",
-            "\\\\",
-            id="bare UNC mount",
-        ),
-        pytest.param(
-            # 2 leading slashes = UNC mount
-            "//mount",
-            "\\\\mount",
-            id="UNC mount",
-        ),
-        pytest.param(
-            # 2 leading slashes = UNC mount
-            "//mount/",
-            "\\\\mount\\",
-            id="UNC mount [trailing]",
-        ),
-        pytest.param(
-            # 3+ leading slashes = root
-            "///",
-            "{WINDOWS}\\Library\\",
-            id="root [leading]",
-        ),
-        pytest.param(
-            # 3+ leading slashes = root path
-            "///root",
-            "{WINDOWS}\\Library\\root",
-            id="root path [leading]",
-        ),
-        pytest.param(
-            # 3+ leading slashes = root
-            "////",
-            "{WINDOWS}\\Library\\",
-            id="root [leading, trailing]",
-        ),
-        pytest.param(
-            # 3+ leading slashes = root path
-            "///root/",
-            "{WINDOWS}\\Library\\root\\",
-            id="root path [leading, trailing]",
-        ),
-        pytest.param(
-            # a normal path
-            "/c/path/to/One",
-            "C:\\path\\to\\One",
-            id="normal path",
-        ),
-        pytest.param(
-            # a normal path
-            "/c//path///to////One",
-            "C:\\path\\to\\One",
-            id="normal path [extra]",
-        ),
-        pytest.param(
-            # a normal path
-            "/c/path/to/One/",
-            "C:\\path\\to\\One\\",
-            id="normal path [trailing]",
-        ),
-        pytest.param(
-            # a normal UNC path
-            "//mount/to/One",
-            "\\\\mount\\to\\One",
-            id="UNC path",
-        ),
-        pytest.param(
-            # a normal UNC path
-            "//mount//to///One",
-            "\\\\mount\\to\\One",
-            id="UNC path [extra]",
-        ),
-        pytest.param(
-            # a normal root path
-            "/path/to/One",
-            "{WINDOWS}\\Library\\path\\to\\One",
-            id="root path",
-        ),
-        pytest.param(
-            # a normal root path
-            "/path//to///One",
-            "{WINDOWS}\\Library\\path\\to\\One",
-            id="root path [extra]",
-        ),
-        pytest.param(
-            # relative path stays relative
-            "relative/path/to/One",
-            "relative\\path\\to\\One",
-            id="relative",
-        ),
-        pytest.param(
-            # relative path stays relative
-            "relative//path///to////One",
-            "relative\\path\\to\\One",
-            id="relative [extra]",
-        ),
-        pytest.param(
-            "/c/path/to/One://path/to/One:/path/to/One:relative/path/to/One",
-            (
-                "C:\\path\\to\\One;"
-                "\\\\path\\to\\One;"
-                "{WINDOWS}\\Library\\path\\to\\One;"
-                "relative\\path\\to\\One"
-            ),
-            id="path;...",
-        ),
-        pytest.param(
-            ["/c/path/to/One"],
-            ("C:\\path\\to\\One",),
-            id="list[path]",
-        ),
-        pytest.param(
-            ("/c/path/to/One", "/c/path/Two", "//mount/Three"),
-            ("C:\\path\\to\\One", "C:\\path\\Two", "\\\\mount\\Three"),
-            id="tuple[path, ...]",
-        ),
-        # XXX Cygwin and MSYS2's cygpath programs are not mutually
-        # aware meaning that MSYS2's cygpath treats
-        # /cygrive/c/here/there as a regular absolute path and returns
-        # {prefix}\Library\cygdrive\c\here\there.  And vice versa.
-        #
-        # cygwin
-        # pytest.param(
-        #     "/cygdrive/c/path/to/One",
-        #     "C:\\path\\to\\One",
-        #     id="Cygwin drive letter path (cygwin)",
-        # ),
-        # pytest.param(
-        #     ["/cygdrive/c/path/to/One"],
-        #     ("C:\\path\\to\\One",),
-        #     id="list[path] (cygwin)",
-        # ),
-        # pytest.param(
-        #     ("/cygdrive/c/path/to/One", "/cygdrive/c/path/Two", "//mount/Three"),
-        #     ("C:\\path\\to\\One", "C:\\path\\Two", "\\\\mount\\Three"),
-        #     id="tuple[path, ...] (cygwin)",
-        # ),
-    ],
-)
-@pytest.mark.parametrize(
-    "unix",
-    [
-        pytest.param(True, id="Unix"),
-        pytest.param(False, id="Windows"),
-    ],
-)
-@pytest.mark.parametrize(
-    "cygpath",
-    [pytest.param(True, id="cygpath"), pytest.param(False, id="fallback")],
-)
-def test_unix_path_to_native(
-    tmp_env: TmpEnvFixture,
-    mocker: MockerFixture,
-    paths: str | Iterable[str] | None,
-    expected: str | tuple[str, ...] | None,
-    unix: bool,
-    cygpath: bool,
-) -> None:
-    windows_prefix = context.target_prefix
-    unix_prefix = native_path_to_unix(windows_prefix)
-
-    def format(path: str) -> str:
-        return path.format(UNIX=unix_prefix, WINDOWS=windows_prefix)
-
-    prefix = unix_prefix if unix else windows_prefix
-    if expected:
-        expected = (
-            tuple(map(format, expected))
-            if isinstance(expected, tuple)
-            else format(expected)
-        )
-
-    if not cygpath:
-        # test without cygpath
-        mocker.patch("subprocess.run", side_effect=FileNotFoundError)
-
-    assert unix_path_to_native(paths, prefix) == expected
 
 
 def test_posix_basic(shell_wrapper_unit: str, monkeypatch: MonkeyPatch):
@@ -3512,6 +3223,9 @@ def test_metavars_force_uppercase(
     "function,raises",
     [
         ("path_identity", TypeError),
+        ("_Cygwin", TypeError),
+        ("native_path_to_unix", TypeError),
+        ("unix_path_to_native", TypeError),
     ],
 )
 def test_deprecations(function: str, raises: type[Exception] | None) -> None:
