+++ tests/test_utils.py
@@ -16,52 +16,10 @@ import pytest
 from conda import CondaError, utils
 from conda.activate import CmdExeActivator, PosixActivator
 from conda.common.compat import on_win
-from conda.common.path import win_path_to_unix
-from conda.testing.helpers import assert_equals
 
 SOME_PREFIX = "/some/prefix"
 SOME_FILES = ["a", "b", "c"]
 log = getLogger(__name__)
-
-
-def test_path_translations():
-    paths = [
-        (
-            r"z:\miniconda\Scripts\pip.exe",
-            "/z/miniconda/Scripts/pip.exe",
-            "/cygdrive/z/miniconda/Scripts/pip.exe",
-        ),
-        (
-            r"z:\miniconda;z:\Documents (x86)\pip.exe;c:\test",
-            "/z/miniconda:/z/Documents (x86)/pip.exe:/c/test",
-            "/cygdrive/z/miniconda:/cygdrive/z/Documents (x86)/pip.exe:/cygdrive/c/test",
-        ),
-        # Failures:
-        # (r"z:\miniconda\Scripts\pip.exe",
-        #  "/z/miniconda/Scripts/pip.exe",
-        #  "/cygdrive/z/miniconda/Scripts/pip.exe"),
-        # ("z:\\miniconda\\",
-        #  "/z/miniconda/",
-        #  "/cygdrive/z/miniconda/"),
-        (
-            "test dummy text /usr/bin;z:\\documents (x86)\\code\\conda\\tests\\envskhkzts\\test1;z:\\documents\\code\\conda\\tests\\envskhkzts\\test1\\cmd more dummy text",
-            "test dummy text /usr/bin:/z/documents (x86)/code/conda/tests/envskhkzts/test1:/z/documents/code/conda/tests/envskhkzts/test1/cmd more dummy text",
-            "test dummy text /usr/bin:/cygdrive/z/documents (x86)/code/conda/tests/envskhkzts/test1:/cygdrive/z/documents/code/conda/tests/envskhkzts/test1/cmd more dummy text",
-        ),
-    ]
-    for windows_path, unix_path, cygwin_path in paths:
-        assert win_path_to_unix(windows_path) == unix_path
-        assert utils.unix_path_to_win(unix_path) == windows_path
-
-        # assert utils.win_path_to_cygwin(windows_path) == cygwin_path
-        # assert utils.cygwin_path_to_win(cygwin_path) == windows_path
-
-
-def test_text_translations():
-    test_win_text = "z:\\msarahan\\code\\conda\\tests\\envsk5_b4i\\test 1"
-    test_unix_text = "/z/msarahan/code/conda/tests/envsk5_b4i/test 1"
-    assert_equals(test_win_text, utils.unix_path_to_win(test_unix_text))
-    assert_equals(test_unix_text, win_path_to_unix(test_win_text))
 
 
 def get_conda_prefixes_on_PATH():
@@ -254,6 +212,8 @@ def test_ensure_dir_errors():
         ("cygwin_path_to_win", TypeError),
         ("translate_stream", TypeError),
         ("path_identity", TypeError),
+        ("unix_path_to_win", TypeError),
+        ("win_path_to_unix", TypeError),
     ],
 )
 def test_deprecations(function: str, raises: type[Exception] | None) -> None:
